Change wctype_t from int to long.
On LP64 this brings us on par with the other libcs where wctype_t is 8
bytes.

Bug: 12875898

+++ libc/include/wchar.h
@@ -65,7 +65,7 @@ enum {
     WC_TYPE_MAX
 };
 
-typedef int wctype_t;
+typedef long wctype_t;
 
 #ifndef WCHAR_MAX
 #define  WCHAR_MAX   INT_MAX
